refactor: update clientType enum
