+++ apps/desktop/src/services/electron-platform-utils.service.ts
@@ -4,7 +4,7 @@ import { I18nService } from "@bitwarden/common/abstractions/i18n.service";
 import { MessagingService } from "@bitwarden/common/abstractions/messaging.service";
 import { PlatformUtilsService } from "@bitwarden/common/abstractions/platformUtils.service";
 import { StateService } from "@bitwarden/common/abstractions/state.service";
-import { ClientType } from "@bitwarden/common/enums/clientType";
+import { ClientType } from "@bitwarden/common/enums/client-type";
 import { DeviceType } from "@bitwarden/common/enums/deviceType";
 
 import { isDev, isMacAppStore } from "../utils";
